ZMETRIC-14: Move MultipleIssueTypeSelector and MultipleWorkflowStatesSelector into subfolder
